fix(core): wire the task relay into the HTTP serve bootstrap path
The ADR-014 relay serving surface (GovernedTaskStore, tasks/* handlers,
ctx.governed_task_store) was wired ONLY in MCPServerFactory. But the HTTP serve
path builds FastMCP directly in server/bootstrap and never used the factory, so
ctx.governed_task_store stayed None and every upstream task handle was rejected
TaskRelayNotSupported regardless of relay_tasks_enabled -- making the #585
activation a no-op for the deployed serving path (issue #591).

Extract the wiring (enable_governed_task_relay + advertise_tasks_capability) into
a shared task_relay_wiring module; MCPServerFactory now delegates to it, and the
HTTP bootstrap calls it after register_all_tools, gated on the
relay_tasks_enabled kill-switch (config key, default True). Behavior through the
factory (and its tests) is unchanged; the HTTP path now activates the relay
identically.
